allow evaluating elisps within haskell functions
This allows for example to send a message to emacs when a async
calculation finished (using unsafePerformIO).

The longterm motivation for this step is to communicate with emacs
within a haskell function, allowing building of something like an Emacs Monad

Note that this is at the moment terribly unsafe.  Nothing prevents
instructions to interleave with other instructions when using
unsafePerformIO causing undefined behaviour.